Refactor: 게시글 상세페이지 좋아요, 공유 컴포넌트 분리
