+++ src/pages/DetailPage/PostArea/PostHeader/Snbs/index.jsx
@@ -1,9 +1,9 @@
 import { useEffect, useState, useRef } from 'react';
-import LikeAndShare from './LikeAndShare';
-import ContentRouter from './ContentRouter';
+import Like from './Like';
+import Share from './Share';
 import styled from 'styled-components';
 
-const Snbs = ({ postId }) => {
+const Snbs = () => {
   const [ScrollY, setScrollY] = useState(0);
   const [ScrollActive, setScrollActive] = useState(false);
   const snbRef = useRef();
@@ -31,8 +31,12 @@ const Snbs = ({ postId }) => {
 
   return (
     <SnbsPositioner ref={snbRef}>
-      <LikeAndShare ScrollActive={ScrollActive} postId={postId} />
-      <ContentRouter ScrollActive={ScrollActive} />
+      <div className={ScrollActive ? 'fixed' : ''}>
+        <SnbContainer>
+          <Like />
+          <Share />
+        </SnbContainer>
+      </div>
     </SnbsPositioner>
   );
 };
@@ -40,11 +44,50 @@ const Snbs = ({ postId }) => {
 const SnbsPositioner = styled.div`
   position: relative;
   margin-top: 2rem;
-  z-index: 15;
+  z-index: 4;
 
   .fixed {
     position: fixed;
     top: 112px;
+  }
+`;
+
+const SnbContainer = styled.div`
+  display: flex;
+  flex-direction: column;
+  -webkit-box-align: center;
+  align-items: center;
+  position: absolute;
+  left: -7rem;
+  padding: 0.5rem;
+  width: 4rem;
+  border: 1px solid var(--border4);
+  border-radius: 2rem;
+  background: var(--bg-element2);
+
+  .icon-circle-wrapper {
+    display: flex;
+    -webkit-box-align: center;
+    align-items: center;
+    -webkit-box-pack: center;
+    justify-content: center;
+    width: 3rem;
+    height: 3rem;
+    border: 1px solid var(--border3);
+    border-radius: 1.5rem;
+    background: var(--bg-element1);
+    color: var(--text3);
+    z-index: 3;
+    cursor: pointer;
+    svg {
+      width: 24px;
+      height: 24px;
+      pointer-events: none;
+    }
+    &:hover {
+      border-color: var(--text1);
+      color: var(--text1);
+    }
   }
 `;
